BE/requirements.txt
mysql-connector-python
python-amazon-simple-product-api